+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.5] - 2021-07-01
 
 - **Added:** `itron::processor::Processor` now implements `TryFrom<usize>`
 - **Added:** `itron::macros`
@@ -47,7 +48,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release.
 
-[Unreleased]: https://github.com/kawadakk/itron-rs/compare/0.1.4...main
+[Unreleased]: https://github.com/kawadakk/itron-rs/compare/0.1.5...main
+[0.1.5]: https://github.com/kawadakk/itron-rs/compare/0.1.4...0.1.5
 [0.1.4]: https://github.com/kawadakk/itron-rs/compare/0.1.3...0.1.4
 [0.1.3]: https://github.com/kawadakk/itron-rs/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/kawadakk/itron-rs/compare/0.1.1...0.1.2